requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pymysql==1.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
authlib==1.3.0
requests==2.31.0
Flask-Mail==0.9.1
chromadb==0.4.22
numpy==1.26.3
Flask-APScheduler==1.13.1
cryptography
